+++ netbox_ssl/api/serializers/__init__.py
@@ -1,10 +1,13 @@
 from .assignments import CertificateAssignmentSerializer
 from .certificate_authorities import CertificateAuthoritySerializer
 from .certificates import CertificateImportSerializer, CertificateSerializer
+from .csr import CertificateSigningRequestSerializer, CSRImportSerializer
 
 __all__ = [
     "CertificateSerializer",
     "CertificateImportSerializer",
     "CertificateAssignmentSerializer",
     "CertificateAuthoritySerializer",
+    "CertificateSigningRequestSerializer",
+    "CSRImportSerializer",
 ]