+++ netbox_ssl/api/urls.py
@@ -10,5 +10,6 @@ router = NetBoxRouter()
 router.register("certificates", views.CertificateViewSet)
 router.register("assignments", views.CertificateAssignmentViewSet)
 router.register("certificate-authorities", views.CertificateAuthorityViewSet)
+router.register("csrs", views.CertificateSigningRequestViewSet)
 
 urlpatterns = router.urls