+++ netbox_ssl/forms/__init__.py
@@ -13,6 +13,12 @@ from .certificates import (
     CertificateForm,
     CertificateImportForm,
 )
+from .csr import (
+    CertificateSigningRequestBulkEditForm,
+    CertificateSigningRequestFilterForm,
+    CertificateSigningRequestForm,
+    CSRImportForm,
+)
 
 __all__ = [
     "CertificateForm",
@@ -24,4 +30,8 @@ __all__ = [
     "CertificateAuthorityForm",
     "CertificateAuthorityFilterForm",
     "CertificateAuthorityBulkEditForm",
+    "CertificateSigningRequestForm",
+    "CertificateSigningRequestFilterForm",
+    "CertificateSigningRequestBulkEditForm",
+    "CSRImportForm",
 ]