+++ netbox_ssl/models/__init__.py
@@ -5,6 +5,7 @@ from .certificate_authorities import (
     CertificateAuthority,
 )
 from .certificates import Certificate, CertificateAlgorithmChoices, CertificateStatusChoices
+from .csr import CertificateSigningRequest, CSRStatusChoices
 
 __all__ = [
     "Certificate",
@@ -14,4 +15,6 @@ __all__ = [
     "CertificateAuthority",
     "CATypeChoices",
     "DEFAULT_CERTIFICATE_AUTHORITIES",
+    "CertificateSigningRequest",
+    "CSRStatusChoices",
 ]