+++ netbox_ssl/urls.py
@@ -60,6 +60,53 @@ urlpatterns = [
         name="certificate_changelog",
         kwargs={"model": models.Certificate},
     ),
+    # CertificateSigningRequest URLs
+    path(
+        "csrs/",
+        views.CertificateSigningRequestListView.as_view(),
+        name="certificatesigningrequest_list",
+    ),
+    path(
+        "csrs/add/",
+        views.CertificateSigningRequestEditView.as_view(),
+        name="certificatesigningrequest_add",
+    ),
+    path(
+        "csrs/import/",
+        views.CSRImportView.as_view(),
+        name="csr_import",
+    ),
+    path(
+        "csrs/edit/",
+        views.CertificateSigningRequestBulkEditView.as_view(),
+        name="certificatesigningrequest_bulk_edit",
+    ),
+    path(
+        "csrs/delete/",
+        views.CertificateSigningRequestBulkDeleteView.as_view(),
+        name="certificatesigningrequest_bulk_delete",
+    ),
+    path(
+        "csrs/<int:pk>/",
+        views.CertificateSigningRequestView.as_view(),
+        name="certificatesigningrequest",
+    ),
+    path(
+        "csrs/<int:pk>/edit/",
+        views.CertificateSigningRequestEditView.as_view(),
+        name="certificatesigningrequest_edit",
+    ),
+    path(
+        "csrs/<int:pk>/delete/",
+        views.CertificateSigningRequestDeleteView.as_view(),
+        name="certificatesigningrequest_delete",
+    ),
+    path(
+        "csrs/<int:pk>/changelog/",
+        ObjectChangeLogView.as_view(),
+        name="certificatesigningrequest_changelog",
+        kwargs={"model": models.CertificateSigningRequest},
+    ),
     # CertificateAuthority URLs
     path(
         "certificate-authorities/",