+++ netbox_ssl/views/__init__.py
@@ -23,6 +23,15 @@ from .certificates import (
     CertificateRenewView,
     CertificateView,
 )
+from .csr import (
+    CertificateSigningRequestBulkDeleteView,
+    CertificateSigningRequestBulkEditView,
+    CertificateSigningRequestDeleteView,
+    CertificateSigningRequestEditView,
+    CertificateSigningRequestListView,
+    CertificateSigningRequestView,
+    CSRImportView,
+)
 
 __all__ = [
     # Certificate views
@@ -34,6 +43,14 @@ __all__ = [
     "CertificateBulkDeleteView",
     "CertificateImportView",
     "CertificateRenewView",
+    # CSR views
+    "CertificateSigningRequestListView",
+    "CertificateSigningRequestView",
+    "CertificateSigningRequestEditView",
+    "CertificateSigningRequestDeleteView",
+    "CertificateSigningRequestBulkEditView",
+    "CertificateSigningRequestBulkDeleteView",
+    "CSRImportView",
     # Certificate Authority views
     "CertificateAuthorityListView",
     "CertificateAuthorityView",